requirements.txt
adjustText==0.8
albumentations==1.3.0
einops==0.6.0
hydra-core==1.3.2
hydra-colorlog==1.2.0
numpy==1.22.2
opencv-python==4.7.0.72
optuna==3.1.0
pandas==1.5.3
python-dotenv==1.0.0
pytorch_lightning==2.0.0
pytest==7.2.2
PyYAML==6.0
scikit-image==0.20.0
scikit-learn==1.2.2
SciencePlots==2.0.1
seaborn==0.12.2
timm==0.6.12
# torch==1.13.1
torchmetrics==0.11.4
# torchvision==0.14.1
wandb==0.14.0
